requirements.txt
ncclient
xmltodict
flask
python-dotenv